Split project into processor & core artifacts (fixes #18)
