+++ gradle.properties
@@ -1,8 +1,10 @@
 VERSION_NAME=1.2-SNAPSHOT
 
 GROUP=com.workday
-POM_ARTIFACT_ID=postman
-POM_NAME=Postman
+POM_ARTIFACT_ID_CORE=postman
+POM_ARTIFACT_ID_PROCESSOR=postman-processor
+POM_NAME_CORE=Postman
+POM_NAME_PROCESSOR=Postman-Processor
 POM_PACKAGING=jar
 
 POM_URL=https://github.com/Workday/postman
