+++ postman-processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,0 +1,1 @@
+com.workday.postman.codegen.PostmanProcessor
