--- postman/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,0 @@
-com.workday.postman.codegen.PostmanProcessor
